ML4HC/Assignment3/requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
ase==3.21.1
asn1crypto==1.0.1
astroid==2.3.1
astropy==3.2.2
astunparse==1.6.3
async-timeout==3.0.1
atom3==0.1.4
atom3-py3==0.1.4.9
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bio==0.4.1
biopython==1.78
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
blessings==1.7
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
cachetools==4.2.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
conda==4.7.12
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
configparser==5.0.2
contextlib2==0.6.0
cryptography==2.7
cv==1.0.0
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
datasets==1.6.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.3
distributed==2.5.2
docker-pycreds==0.4.0
docutils==0.15.2
easy-parallel==0.1.6
easy-parallel-py3==0.1.6.4
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
flatbuffers==1.12
fsspec==2021.5.0
future==0.17.1
gast==0.3.3
gevent==1.4.0
gitdb==4.0.7
GitPython==3.1.14
glob2==0.7
gmpy2==2.0.8
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googledrivedownloader==0.4
gpustat==0.6.0
greenlet==0.4.15
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
huggingface-hub==0.0.8
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jeepney==0.4.1
Jinja2==2.10.3
joblib==0.13.2
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras-Preprocessing==1.1.2
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
mpmath==1.1.0
msgpack==0.6.1
multidict==5.1.0
multipledispatch==0.6.0
multiprocess==0.70.11.1
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nglview==2.7.7
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.19.5
numpydoc==0.9.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
openpyxl==3.0.0
opt-einsum==3.3.0
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
pathos==0.2.7
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
pox==0.2.9
ppft==1.6.6.3
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.15.6
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyDeprecate==0.3.0
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.4.2
pyodbc==4.0.27
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.0
python-louvain==0.15
pytorch-lightning==1.3.4
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
rdflib==5.0.0
regex==2021.4.4
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.7.2
ruamel-yaml==0.15.46
sacremoses==0.0.45
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
seaborn==0.11.1
SecretStorage==3.1.1
Send2Trash==1.5.0
sentry-sdk==1.1.0
shortuuid==1.0.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
smmap==4.0.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.9
statsmodels==0.10.1
subprocess32==3.5.4
sympy==1.4
tables==3.5.2
tblib==1.4.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tokenizers==0.10.2
toolz==0.10.0
torch==1.8.1+cu111
torch-cluster==1.5.9
torch-geometric==1.6.3
torch-scatter==2.0.6
torch-sparse==0.6.9
torch-spline-conv==1.2.1
torchaudio==0.8.1
torchmetrics==0.3.2
torchvision==0.9.1+cu111
tornado==6.0.3
tqdm==4.61.0
traitlets==4.3.3
transformers==4.5.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.24.2
wandb==0.10.32
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwt==1.3.0
xxhash==2.0.2
yarl==1.6.3
zict==1.0.0
zipp==0.6.0
